invoice-worker/requirements.txt
python-dotenv==1.0.0
psycopg2-binary==2.9.9
requests==2.31.0
azure-storage-queue==12.9.0
azure-storage-blob==12.19.0
reportlab==4.0.7

# Testing
pytest==7.4.3
